requirements.txt
# Discord & Telegram
discord.py>=2.3.0
python-telegram-bot>=20.0

# AI APIs
google-generativeai>=0.3.0  # Google Gemini (recommended)
openai>=1.0.0               # OpenAI GPT (optional)
anthropic>=0.18.0           # Anthropic Claude (optional)

# Voice - Cloud APIs
elevenlabs>=0.2.0
azure-cognitiveservices-speech>=1.35.0

# Voice - Self-hosted (Toucan TTS)
# Install separately: git clone https://github.com/DigitalPhonetics/IMS-Toucan && cd IMS-Toucan && pip install -e .

# Audio processing
pydub>=0.25.0

# Anki integration
genanki>=0.13.0
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Scheduling (optional - for automation)
apscheduler>=3.10.0
